+++ tests/test_simple_fields.py
@@ -17,10 +17,10 @@ class SimpleFieldsTest(unittest.TestCase):
                       '2017-07-26T08:31:05', '2017-07-26T08:31:05Z', '2017-07-26 08:31',
                       '2017-07-26T13:31:05+05', '2017-07-26 13:31:05+0500'):
             dt = f.to_python(value, pytz.utc)
-            self.assertEquals(dt.tzinfo, pytz.utc)
+            self.assertEqual(dt.tzinfo, pytz.utc)
             # Verify that conversion to and from db string does not change value
             dt2 = f.to_python(f.to_db_string(dt, quote=False), pytz.utc)
-            self.assertEquals(dt, dt2)
+            self.assertEqual(dt, dt2)
         # Invalid values
         for value in ('nope', '21/7/1999', 0.5,
                       '2017-01 15:06:00', '2017-01-01X15:06:00', '2017-13-01T15:06:00'):
@@ -33,10 +33,10 @@ class SimpleFieldsTest(unittest.TestCase):
         # Valid values
         for value in (datetime(1970, 1, 1), epoch, '1970-01-01', '0000-00-00', 0):
             d = f.to_python(value, pytz.utc)
-            self.assertEquals(d, epoch)
+            self.assertEqual(d, epoch)
             # Verify that conversion to and from db string does not change value
             d2 = f.to_python(f.to_db_string(d, quote=False), pytz.utc)
-            self.assertEquals(d, d2)
+            self.assertEqual(d, d2)
         # Invalid values
         for value in ('nope', '21/7/1999', 0.5):
             with self.assertRaises(ValueError):
@@ -50,7 +50,7 @@ class SimpleFieldsTest(unittest.TestCase):
         # Verify that conversion of timezone-aware datetime is correct
         f = DateField()
         dt = datetime(2017, 10, 5, tzinfo=pytz.timezone('Asia/Jerusalem'))
-        self.assertEquals(f.to_python(dt, pytz.utc), date(2017, 10, 4))
+        self.assertEqual(f.to_python(dt, pytz.utc), date(2017, 10, 4))
 
     def test_datetime_field_timezone(self):
         # Verify that conversion of timezone-aware datetime is correct
@@ -63,13 +63,13 @@ class SimpleFieldsTest(unittest.TestCase):
                 '2017-07-26 11:31:05+0300',
                 '2017-07-26T03:31:05-0500',
         ):
-            self.assertEquals(f.to_python(value, pytz.utc), utc_value)
+            self.assertEqual(f.to_python(value, pytz.utc), utc_value)
 
     def test_uint8_field(self):
         f = UInt8Field()
         # Valid values
         for value in (17, '17', 17.0):
-            self.assertEquals(f.to_python(value, pytz.utc), 17)
+            self.assertEqual(f.to_python(value, pytz.utc), 17)
         # Invalid values
         for value in ('nope', date.today()):
             with self.assertRaises(ValueError):