+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+v1.0.2
+----------
+- Include alias and materialized fields in queryset results
+- Check for database existence, to allow delayed creation
+- Added `Database.does_table_exist` method
+- Support for `IS NULL` and `IS NOT NULL` in querysets (kalombos)
+
 v1.0.1
 ------
 - NullableField: take extra_null_values into account in `validate` and `to_python`